Fix critical Rust compilation errors and enable native module builds

+++ src/inventory.rs
@@ -25,8 +25,8 @@ pub struct EOQCalculation {
     pub total_annual_cost: f64,        // Keep f64 for precise financial calculations
     pub ordering_cost_component: f64,  // Keep f64 for precise financial calculations
     pub holding_cost_component: f64,   // Keep f64 for precise financial calculations
-    pub order_frequency: f32,          // Changed to f32 for frequency metrics
-    pub optimal_cycle_time_days: f32,  // Changed to f32 for time calculations
+    pub order_frequency: f64,          // Changed to f64 for frequency metrics
+    pub optimal_cycle_time_days: f64,  // Changed to f64 for time calculations
 }
 
 #[derive(Serialize, Deserialize)]
@@ -34,9 +34,9 @@ pub struct EOQCalculation {
 pub struct SafetyStockCalculation {
     pub item_id: String,
     pub safety_stock: i32,           // Keep as integer for exact quantities
-    pub service_level: f32,          // Changed to f32 for percentage metrics
-    pub demand_variability: f32,     // Changed to f32 for statistical metrics
-    pub lead_time_variability: f32,  // Changed to f32 for statistical metrics
+    pub service_level: f64,          // Changed to f64 for percentage metrics
+    pub demand_variability: f64,     // Changed to f64 for statistical metrics
+    pub lead_time_variability: f64,  // Changed to f64 for statistical metrics
     pub reorder_point: i32,          // Keep as integer for exact quantities
 }
 
@@ -45,8 +45,8 @@ pub struct SafetyStockCalculation {
 pub struct ABCAnalysis {
     pub item_id: String,
     pub annual_usage_value: f64,      // Keep f64 for precise financial calculations
-    pub percentage_of_total_value: f32, // Changed to f32 for percentage metrics
-    pub cumulative_percentage: f32,     // Changed to f32 for percentage metrics
+    pub percentage_of_total_value: f64, // Changed to f64 for percentage metrics
+    pub cumulative_percentage: f64,     // Changed to f64 for percentage metrics
     pub abc_class: String,
     pub management_priority: String,
 }
